DCWL-2003-v2 renamed specs to match services
