+++ test/unit/services/filters/VersionResponseFilterServiceSpec.scala
@@ -14,22 +14,23 @@
  * limitations under the License.
  */
 
-package unit.services
+package unit.services.filters
 
 import org.scalatest.Assertion
 import play.api.test.Helpers
 import uk.gov.hmrc.customs.api.common.xml.ValidateXmlAgainstSchema
+import uk.gov.hmrc.customs.api.common.xml.ValidateXmlAgainstSchema._
+import uk.gov.hmrc.customs.declarations.information.services.filters.VersionResponseFilterService
 import util.UnitSpec
 import util.VersionTestXMLData.{defaultDateTime, generateDeclarationResponseContainingAllOptionalElements, generateDeclarationVersionResponse, validBackendVersionResponse}
+
 import javax.xml.validation.Schema
 import scala.concurrent.ExecutionContext
 import scala.util.Try
 import scala.xml._
 import scala.xml.transform.{RewriteRule, RuleTransformer}
-import ValidateXmlAgainstSchema._
-import uk.gov.hmrc.customs.declarations.information.services.filters.VersionResponseFilterService
 
-class DeclarationVersionResponseFilterServiceSpec extends UnitSpec {
+class VersionResponseFilterServiceSpec extends UnitSpec {
   implicit val ec: ExecutionContext = Helpers.stubControllerComponents().executionContext
   val schemaFile: Try[Schema] = getSchema("/api/conf/1.0/schemas/wco/declaration/DeclarationInformationRetrievalVersionResponse.xsd")
   val xmlValidationService: ValidateXmlAgainstSchema = new ValidateXmlAgainstSchema(schemaFile.get)
